Moved get_transmission to arflux.py
