+++ pypit/arflux.py
@@ -497,3 +497,46 @@ def lrisr_800_10000_r(wavelength):
     resr = resr/wavelength*3.E5 # km s^-1
 
     return np.array(resr)
+
+def get_transmission(atm_file, data):
+    """ Convolve the atmospheric absorption model to the
+    resolution and wavelengthg gird of the data.
+    Parameters
+    ----------
+    atm_file: string
+        filename passed through script call
+    data: XSpectrum1d object
+    Returns
+    -------
+    model: 2D ndarray
+    """
+    from pkg_resources import resource_filename
+    from pypit import arutils
+
+    floc  = ('data/extinction/{}'.format(atm_file))
+    fname = resource_filename('pypit', floc)
+    tmp   = np.loadtxt(fname)
+
+    # Convert transmission spectrum to angstroms
+    # and clip it to wavelength range of data
+    i = ((tmp[:,0]*10 >= min(data.wavelength.value)) &
+         (tmp[:,0]*10 <= max(data.wavelength.value)))
+    trans = XSpectrum1D.from_tuple((tmp[:,0][i]*10, tmp[:,1][i]))
+
+    # Convolve the atm. transsissiom spectrum to LRIS_R resolution
+    res         = lrisr_800_10000_r(trans.wavelength.value)
+    fwhm_pix    = arutils.get_fwhm_pix(trans, inval=20000., outval=res)
+
+    smooth_spec = np.zeros((len(trans.wavelength.value)))
+    for i, fp in enumerate(fwhm_pix):
+        tmp = trans.gauss_smooth(fp)
+        smooth_spec[i] = tmp.flux.value[i]
+
+    # Get model transmission on same wavelength grid
+    # as data
+    model = np.zeros((len(data.wavelength.value), 2))
+    model[:,0] = data.wavelength.value
+    model[:,1] = np.interp(data.wavelength.value,
+                           trans.wavelength.value, smooth_spec)
+
+    return model
